fix: verify email build issue
